fix: remove social preview cache flow
