+++ api/services/articleMetadata.ts
@@ -1,0 +1,9 @@
+export type ArticleMetadata = {
+  title: string
+  summary: string
+  image: string
+  author: string
+  published?: number
+  tags?: string[]
+  imageAlt?: string
+}
