+++ vercel.json
@@ -3,9 +3,6 @@
   "functions": {
     "api/article-og.ts": {
       "maxDuration": 10
-    },
-    "api/article-og-refresh.ts": {
-      "maxDuration": 10
     }
   },
   "rewrites": [
@@ -26,4 +23,3 @@
     }
   ]
 }
-
